Removed global.unusedSpawns, global.uniqueSpawns tables.
global.allSpawns has a .createdFor field with the name of the controlling player, and a .used field indicating spawn is in use.Spawns are assumed to be on the standard game surface.

global.playerSpawns has surface name (for future use), and seq of the slot originally assigned.

New command /spawns shows status of the spawns.

Implement an AddSpawn() function to add spawns dynamically as the game progresses.

If the number of available spawns is zero, don't show the solo spawn button.